+++ canard.h
@@ -203,7 +203,7 @@ struct CanardRxState
 
     CanardBufferBlock* buffer_blocks;
 
-    uint64_t timestamp_usec;
+    uint32_t timestamp_usec;
 
     const uint32_t dtid_tt_snid_dnid;
 
@@ -247,7 +247,7 @@ struct CanardRxTransfer
     /**
      * Timestamp at which the first frame of this transfer was received.
      */
-    uint64_t timestamp_usec;
+    uint32_t timestamp_usec;
 
     /**
      * Payload is scattered across three storages:
@@ -391,7 +391,7 @@ void canardPopTxQueue(CanardInstance* ins);
  */
 void canardHandleRxFrame(CanardInstance* ins,
                          const CanardCANFrame* frame,
-                         uint64_t timestamp_usec);
+                         uint32_t timestamp_usec);
 
 /**
  * Traverses the list of transfers and removes those that were last updated more than timeout_usec microseconds ago.
@@ -399,7 +399,7 @@ void canardHandleRxFrame(CanardInstance* ins,
  * Also refer to the constant CANARD_RECOMMENDED_STALE_TRANSFER_CLEANUP_INTERVAL_USEC.
  */
 void canardCleanupStaleTransfers(CanardInstance* ins,
-                                 uint64_t current_time_usec);
+                                 uint32_t current_time_usec);
 
 /**
  * This function can be used to extract values from received UAVCAN transfers. It decodes a scalar value -
